% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mictools.R
\name{mic_strength}
\alias{mic_strength}
\title{This function compute the mic strenght based on statistical significance of the pvalue computed on null distribution of TIC.}
\usage{
mic_strength(x, pval, alpha = NULL, C = 5, pthr = 0.05,
  pval.col = 1:3)
}
\arguments{
\item{x}{a numeric vector (of size \emph{n}), matrix or data frame (which is coerced to matrix).}

\item{pval}{a data.frame with pvalues for each pair of association of the x input matrix. It should contain two colums with 
the indices of the computed association according to the x input matrix}

\item{alpha}{an optional number of cells allowed in the \emph{X-by-Y} search-grid. Default value is 0.6 (see Details).}

\item{C}{an optional number determining the starting point of the 
\emph{X-by-Y} search-grid. When trying to partition the \emph{x}-axis into 
\emph{X} columns, the algorithm will start with at most \code{C}\emph{X} 
\emph{clumps}. Default value is 15 (see Details).}

\item{pthr}{threshold on pvalue for measure to consider for computing mic_e}

\item{pval.col}{a vector of integer or character of length 3 indicating the columns of the pval data.frame with the index of the 
matrix x of the association computed. The first index refers to the column containing the p.values, while the remaining 2 indices
refers to x variables.}
}
\description{
This function compute the mic strenght based on statistical significance of the pvalue computed on null distribution of TIC.
}
