% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mictools_null}
\alias{mictools_null}
\title{Function to compute mictools null distribution}
\usage{
mictools_null(x, alpha = 9, C = 5, nperm = 250000L, seed = 0L)
}
\arguments{
\item{x}{Numeric Vector}

\item{alpha}{alpha parameter for the mine statistic}

\item{C}{c parameter for the mine statistic}

\item{nperm}{numper of permutation}

\item{seed}{character which measure to return}
}
\description{
Function to compute mictools null distribution
}
