% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mictools_pstats}
\alias{mictools_pstats}
\title{Function to compute the pvalue for the mictools pipeline}
\usage{
mictools_pstats(x, alpha = 0.6, C = 15L, est = "mic_approx")
}
\arguments{
\item{x}{Numeric Vector}

\item{alpha}{alpha parameter for the mine statistic}

\item{C}{c parameter for the mine statistic}

\item{est}{estimation parameter for the mine statistic}
}
\description{
Function to compute the pvalue for the mictools pipeline
}
