% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mine_compute}
\alias{mine_compute}
\title{Function to compute one statistic at time}
\usage{
mine_compute(x, y, alpha = 0.6, C = 15, est = "mic_approx",
  measure = 1L, eps = 0, p = -1, norm = FALSE)
}
\arguments{
\item{x}{Numeric Vector}

\item{y}{Numeric Vector}

\item{alpha}{alpha parameter for the mine statistic}

\item{C}{c parameter for the mine statistic}

\item{est}{estimation parameter for the mine statistic}

\item{measure}{character which measure to return}

\item{eps}{eps value for MCN statistic}

\item{p}{probability for the generalized mic}

\item{norm}{boolean if require normalization between 0 and 1 for the tic statistic}
}
\description{
Function to compute one statistic at time
}
