\encoding{UTF-8}
\name{Spellman}
\docType{data}
\alias{Spellman}
\alias{spellman}
\title{CDC15 Yeast Gene Expression Dataset}
\description{
The Spellman dataset provides the gene expression data measured (on a custom platform) 
in \emph{Saccharomyces cerevisiae} cell cultures that have been synchronized at 
different points of the cell cycle by using a temperature-sensitive mutation 
(\emph{cdc15-2}), which arrestes cells late in mitosis at the restrictive 
temperature (it can cause heat-shock). 

}
\usage{Spellman}
\format{23 rows x 4382 columns: 4381 transcripts (columns 2:4382)
  measured at 23 timepoints (column 1).}
\source{The original data were published by Spellman and colleagues in Mol. 
Biol. Cell (1998) as the Botstein dataset. Here we include the version of the 
dataset as processed by Reshef and colleagues for the MINE statistics original 
article published in Science (2011) (details are provided in the supplementary material).}
\references{

  D. Reshef, Y. Reshef, H. Finucane, S. Grossman, G. McVean, P. Turnbaugh, 
  E. Lander, M. Mitzenmacher, P. Sabeti. (2011)
  \emph{Detecting novel associations in large datasets}. 
  Science 334, 6062  
  (\url{http://www.exploredata.net}).

  P. T. Spellman, G. Sherlock, M. Q. Zhang, V. R. Iyer, K. Anders, M. B. Eisen,
  P. O. Brown, D. Botstein, B. Futcher. (1998)
  \emph{Comprehensive Identification of Cell Cycle–regulated Genes of the Yeast 
  Saccharomyces cerevisiae by Microarray Hybridization}.
  Mol. Biol. Cell, 9:12 3273--3297.

}
\keyword{datasets}
