% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mictools.R
\name{mictools}
\alias{mictools}
\title{Function for mictools pipeline}
\usage{
mictools(x, alpha = 9, C = 5, seed = 0, nperm = 250000)
}
\arguments{
\item{nperm}{integer, number of permutation to perform}
}
\description{
Function for mictools pipeline
}
