% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mictools_null}
\alias{mictools_null}
\title{This set of functions are helper function to compute null distribution of the \code{tic_e} and 
\code{tic_e} observed distribution from a matrix}
\usage{
mictools_null(x, alpha = 9, C = 5, nperm = 200000L, seed = 0L)
}
\arguments{
\item{x}{matrix N x M with M variables and N samples}

\item{alpha}{numeric value representing parameter for the mine statistic see \code{\link[minerva]{mine}}}

\item{C}{c parameter for the mine statistic see \code{\link[minerva]{mine}}}

\item{nperm}{numper of permutation}

\item{seed}{integer to set the starting seed for random number generation (for reproducibility).}
}
\value{
It returns a vector of \code{nperm} \code{tic_e} values.
}
\description{
This set of functions are helper function to compute null distribution of the \code{tic_e} and 
\code{tic_e} observed distribution from a matrix
}
\section{Functions}{
\itemize{
\item \code{mictools_null}: compute the \code{tic_e} null distribution
}}

\seealso{
\code{\link[minerva]{mictools}}
}
